# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Cars Binar Rent Documentation",
    "description": "Dokumentasi API Rental obil",
    "contact": {
      "email": "nexsel@gmail.com"
    },
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:5000/api/v1"
    },
    {
      "url": "http://localhost:3000/api/v1"
    }
  ],
...
  "paths": {
...
    "/cars": {
      "get": {
        "tags": ["cars"],
        "summary": "Get cars lists",
        "description": "Get cars Lists",
        "operationId": "getcars",
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "required": [
                    "name",
                    "category",
                    "price",
                    "start_rent",
                    "finish_rent",
                    "image_url"
                  ],
                  "items": {
                    "type": "object",
                    "properties": {
                      "name": {
                        "type": "string",
                        "example": "Car A"
                      },
                      "category": {
                        "type": "string",
                        "example": "Shakespear"
                      },
                      "price": {
                        "type": "number",
                        "example": 50000
                      },
                      "start_rent": {
                        "type": "string",
                        "example": "2024-05-31T17:00:00.000Z"
                      },
                      "finish_rent": {
                        "type": "string",
                        "example": "2024-06-09T17:00:00.000Z"
                      },
                      "image_url": {
                        "type": "string",
                        "example": "http://cloudinary.com/213125123"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "Error!"
                    },
                    "code": {
                      "type": "integer",
                      "example": 500
                    },
                    "message": {
                      "type": "string",
                      "example": "Internal Server Error!"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["cars"],
        "summary": "Add a new cars",
        "description": "Add a new cars",
        "operationId": "addcars",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": [
                  "name",
                  "category",
                  "price",
                  "start_rent",
                  "finish_rent",
                  "image_url"
                ],
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Car A"
                  },
                  "size": {
                    "type": "string",
                    "enum": ["small", "medium", "large"],
                    "example": "medium"
                  },
                  "category": {
                    "type": "string",
                    "example": "Shakespear"
                  },
                  "price": {
                    "type": "number",
                    "example": 50000
                  },
                  "start_rent": {
                    "type": "string",
                    "example": "2024-05-31T17:00:00.000Z"
                  },
                  "finish_rent": {
                    "type": "string",
                    "example": "2024-06-09T17:00:00.000Z"
                  },
                  "image_url": {
                    "type": "string",
                    "format": "binary",
                    "example": "http://www.africau.edu/images/default/sample.pdf"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer",
                            "example": 1
                          },
                          "size": {
                            "type": "string",
                            "enum": ["small", "medium", "large"],
                            "example": "large"
                          },
                          "name": {
                            "type": "string",
                            "example": "Car A"
                          },
                          "price": {
                            "type": "number",
                            "example": 50000
                          },
                          "category": {
                            "type": "string",
                            "example": "Shakespear"
                          },
                          "start_rent": {
                            "type": "string",
                            "format": "date-time",
                            "example": "2024-05-31T17:00:00.000Z"
                          },
                          "finish_rent": {
                            "type": "string",
                            "format": "date-time",
                            "example": "2024-06-09T17:00:00.000Z"
                          },
                          "image_url": {
                            "type": "string",
                            "format": "uri",
                            "example": "http://res.cloudinary.com/dcpewaqiq/image/upload/v1717339636/hwciyatpenzbdi1sgyck.jpg"
                          },
                          "created_by": {
                            "type": "string",
                            "example": "test"
                          },
                          "updated_by": {
                            "type": "string",
                            "example": "test"
                          },
                          "created_at": {
                            "type": "string",
                            "format": "date-time",
                            "example": "2024-06-02T14:47:17.234Z"
                          },
                          "updated_at": {
                            "type": "string",
                            "format": "date-time",
                            "example": "2024-06-02T14:47:17.234Z"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Email sudah terdaftar",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "Fail!"
                    },
                    "message": {
                      "type": "string",
                      "example": "Email sudah terdaftar!"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "auth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "auth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}